Движок обновлен до v8.1.3 Теперь сервер обнуляет кол-во мертвых после рестарта Исправлен баг с двойным уроном

+++ src/game/client/person/player/Bullet.java
@@ -60,6 +60,8 @@ public class Bullet extends Obj implements Collision.CollisionListener{
 
 	@Override
 	public void collision(Obj obj) {
+		if (destroy) return;
+
 		if (obj.getClass().equals(Wall.class) ||
 			obj.getClass().equals(Border.class)){
 				destroyBullet();
